+++ nautilus_trader/persistence/external/util.py
@@ -6,6 +6,10 @@ from typing import Optional
 
 import pandas as pd
 
+from nautilus_trader.core.nautilus_pyo3.persistence import ParquetType
+from nautilus_trader.model.data.tick import QuoteTick
+from nautilus_trader.model.data.tick import TradeTick
+
 
 class Singleton(type):
     """
@@ -103,3 +107,12 @@ def parse_filename_start(fn: str) -> Optional[tuple[str, pd.Timestamp]]:
 
     start = pd.Timestamp(start)
     return instrument_id, start
+
+
+def py_type_to_parquet_type(cls: type) -> ParquetType:
+    if cls == QuoteTick:
+        return ParquetType.QuoteTick
+    elif cls == TradeTick:
+        return ParquetType.TradeTick
+    else:
+        raise RuntimeError(f"Type {cls} not supported as a `ParquetType` yet.")
